add use of AE_SUPPORT_UDP
