+++ aether/CMakeLists.txt
@@ -166,20 +166,20 @@ list(APPEND serial_ports_srcs
             )
 
 list(APPEND transport_srcs
-            "transport/low_level/sockets/unix_socket.cpp"
-            "transport/low_level/sockets/unix_tcp_socket.cpp"
-            "transport/low_level/sockets/unix_udp_socket.cpp"
-            "transport/low_level/sockets/lwip_socket.cpp"
-            "transport/low_level/sockets/lwip_tcp_socket.cpp"
-            "transport/low_level/sockets/lwip_udp_socket.cpp"
-            "transport/low_level/sockets/win_socket.cpp"
-            "transport/low_level/sockets/win_tcp_socket.cpp"
-            "transport/low_level/sockets/win_udp_socket.cpp"
+            "transport/system_sockets/sockets/unix_socket.cpp"
+            "transport/system_sockets/sockets/unix_tcp_socket.cpp"
+            "transport/system_sockets/sockets/unix_udp_socket.cpp"
+            "transport/system_sockets/sockets/lwip_socket.cpp"
+            "transport/system_sockets/sockets/lwip_tcp_socket.cpp"
+            "transport/system_sockets/sockets/lwip_udp_socket.cpp"
+            "transport/system_sockets/sockets/win_socket.cpp"
+            "transport/system_sockets/sockets/win_tcp_socket.cpp"
+            "transport/system_sockets/sockets/win_udp_socket.cpp"
             "transport/modems/modem_transport.cpp"
             "transport/socket_packet_send_action.cpp"
             "transport/data_packet_collector.cpp"
-            "transport/low_level/tcp/tcp.cpp"
-            "transport/low_level/udp/udp.cpp"
+            "transport/system_sockets/tcp/tcp.cpp"
+            "transport/system_sockets/udp/udp.cpp"
             "transport/modems/modem_transport.cpp"
             "transport/lora_modules/lora_module_transport.cpp"
 )
