+++ aether/channels/modem_channel.cpp
@@ -145,20 +145,17 @@ ModemChannel::ModemChannel(ObjPtr<Aether> aether,
       aether_{std::move(aether)},
       access_point_{std::move(access_point)} {
   // fill transport properties
+  transport_properties_.max_packet_size = 1024;
+  transport_properties_.rec_packet_size = 1024;
   auto protocol = std::visit([](auto&& adr) { return adr.protocol; }, address);
-
   switch (protocol) {
     case Protocol::kTcp: {
       transport_properties_.connection_type = ConnectionType::kConnectionFull;
-      transport_properties_.max_packet_size = 1024;
-      transport_properties_.rec_packet_size = 1024;
       transport_properties_.reliability = Reliability::kReliable;
       break;
     }
     case Protocol::kUdp: {
       transport_properties_.connection_type = ConnectionType::kConnectionLess;
-      transport_properties_.max_packet_size = 1024;
-      transport_properties_.rec_packet_size = 1024;
       transport_properties_.reliability = Reliability::kUnreliable;
       break;
     }
