+++ aether/transport/system_sockets/sockets/isocket.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef AETHER_TRANSPORT_LOW_LEVEL_SOCKETS_ISOCKET_H_
-#define AETHER_TRANSPORT_LOW_LEVEL_SOCKETS_ISOCKET_H_
+#ifndef AETHER_TRANSPORT_SYSTEM_SOCKETS_SOCKETS_ISOCKET_H_
+#define AETHER_TRANSPORT_SYSTEM_SOCKETS_SOCKETS_ISOCKET_H_
 
 #include <cstdint>
 #include <cstddef>
@@ -88,4 +88,4 @@ class ISocket {
 };
 }  // namespace ae
 
-#endif  // AETHER_TRANSPORT_LOW_LEVEL_SOCKETS_ISOCKET_H_
+#endif  // AETHER_TRANSPORT_SYSTEM_SOCKETS_SOCKETS_ISOCKET_H_
