+++ aether/transport/system_sockets/sockets/unix_socket.h
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-#ifndef AETHER_TRANSPORT_LOW_LEVEL_SOCKETS_UNIX_SOCKET_H_
-#define AETHER_TRANSPORT_LOW_LEVEL_SOCKETS_UNIX_SOCKET_H_
+#ifndef AETHER_TRANSPORT_SYSTEM_SOCKETS_SOCKETS_UNIX_SOCKET_H_
+#define AETHER_TRANSPORT_SYSTEM_SOCKETS_SOCKETS_UNIX_SOCKET_H_
 
 #if defined(__linux__) || defined(__unix__) || defined(__APPLE__) || \
     defined(__FreeBSD__)
 
 #  define UNIX_SOCKET_ENABLED 1
 
-#  include "aether/transport/low_level/sockets/isocket.h"
+#  include "aether/transport/system_sockets/sockets/isocket.h"
 
 namespace ae {
 /**
@@ -50,4 +50,4 @@ class UnixSocket : public ISocket {
 
 #endif
 
-#endif  // AETHER_TRANSPORT_LOW_LEVEL_SOCKETS_UNIX_SOCKET_H_
+#endif  // AETHER_TRANSPORT_SYSTEM_SOCKETS_SOCKETS_UNIX_SOCKET_H_
