+++ aether/transport/system_sockets/sockets/win_socket.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef AETHER_TRANSPORT_LOW_LEVEL_SOCKETS_WIN_SOCKET_H_
-#define AETHER_TRANSPORT_LOW_LEVEL_SOCKETS_WIN_SOCKET_H_
+#ifndef AETHER_TRANSPORT_SYSTEM_SOCKETS_SOCKETS_WIN_SOCKET_H_
+#define AETHER_TRANSPORT_SYSTEM_SOCKETS_SOCKETS_WIN_SOCKET_H_
 
 #if defined _WIN32
 #  define WIN_SOCKET_ENABLED 1
@@ -27,7 +27,7 @@
 #  include "aether/types/data_buffer.h"
 #  include "aether/socket_initializer.h"
 
-#  include "aether/transport/low_level/sockets/isocket.h"
+#  include "aether/transport/system_sockets/sockets/isocket.h"
 
 namespace ae {
 class WinSocket : public ISocket {
@@ -64,4 +64,4 @@ class WinSocket : public ISocket {
 }  // namespace ae
 #endif
 
-#endif  // AETHER_TRANSPORT_LOW_LEVEL_SOCKETS_WIN_SOCKET_H_
+#endif  // AETHER_TRANSPORT_SYSTEM_SOCKETS_SOCKETS_WIN_SOCKET_H_
