+++ aether/transport/system_sockets/sockets/win_udp_socket.cpp
@@ -14,22 +14,19 @@
  * limitations under the License.
  */
 
-#include "aether/transport/low_level/sockets/win_tcp_socket.h"
-
-#if defined WIN_SOCKET_ENABLED
+#include "aether/transport/system_sockets/sockets/win_udp_socket.h"
+#if AE_SUPPORT_UDP && defined WIN_SOCKET_ENABLED
 
 #  include <winsock2.h>
 #  include <ws2def.h>
 #  include <ws2ipdef.h>
-#  include <mswsock.h>
 
 #  include "aether/misc/defer.h"
 
 #  include "aether/tele/tele.h"
 
 namespace ae {
-
-namespace win_socket_internal {
+namespace win_udp_socket_internal {
 struct SockAddr {
   sockaddr* addr() { return reinterpret_cast<sockaddr*>(&data); }
 
@@ -87,14 +84,13 @@ SockAddr GetSockAddr(IpAddressPort const& ip_address_port) {
   }
   return {};
 }
-}  // namespace win_socket_internal
+}  // namespace win_udp_socket_internal
 
-WinTcpSocket::WinTcpSocket()
-    : WinSocket{1500}, conn_overlapped_{{}, EventType::kRead} {
+WinUdpSocket::WinUdpSocket() : WinSocket{1200} {
   bool created = false;
 
   // ::socket() sets WSA_FLAG_OVERLAPPED by default that allows us to use iocp
-  auto sock = ::socket(AF_INET, SOCK_STREAM, 0);
+  auto sock = ::socket(AF_INET, SOCK_DGRAM, 0);
   if (sock == INVALID_SOCKET) {
     AE_TELED_ERROR("Got socket creation error {}", WSAGetLastError());
     return;
@@ -105,14 +101,6 @@ WinTcpSocket::WinTcpSocket()
       ::closesocket(sock);
     }
   };
-
-  int on = 1;
-  if (auto res = ::setsockopt(sock, IPPROTO_TCP, TCP_NODELAY,
-                              reinterpret_cast<char const*>(&on), sizeof(on));
-      res == SOCKET_ERROR) {
-    AE_TELED_ERROR("Socket connect error {}", WSAGetLastError());
-    return;
-  }
 
   // make socket non-blocking
   u_long nonblocking = 1;
@@ -126,110 +114,49 @@ WinTcpSocket::WinTcpSocket()
   socket_ = sock;
 }
 
-WinTcpSocket::~WinTcpSocket() { Disconnect(); }
+WinUdpSocket::~WinUdpSocket() { Disconnect(); }
 
-WinTcpSocket::ConnectionState WinTcpSocket::Connect(
+WinUdpSocket::ConnectionState WinUdpSocket::Connect(
     IpAddressPort const& destination) {
-  assert(socket_ != INVALID_SOCKET);
-
-  auto connect_ex = GetConnectEx();
-  if (connect_ex == nullptr) {
+  auto addr = win_udp_socket_internal::GetSockAddr(destination);
+  auto res = connect(socket_, reinterpret_cast<sockaddr*>(&addr), sizeof(addr));
+  if (res == SOCKET_ERROR) {
+    AE_TELED_ERROR("Socket connect error {}", WSAGetLastError());
     return ConnectionState::kConnectionFailed;
   }
 
-  // Bind socket first (required for proper IOCP connection handling)
-  sockaddr_in localAddr = {};
-  localAddr.sin_family = AF_INET;
-  localAddr.sin_addr.s_addr = INADDR_ANY;
-  localAddr.sin_port = 0;  // Let system choose port
-
-  if (bind(socket_, reinterpret_cast<sockaddr*>(&localAddr),
-           sizeof(localAddr)) == SOCKET_ERROR) {
-    AE_TELED_ERROR("Socket bind error {}", WSAGetLastError());
+  if (!RequestRecv()) {
+    AE_TELED_ERROR("Request receive failed!");
     return ConnectionState::kConnectionFailed;
-  }
-
-  auto addr = win_socket_internal::GetSockAddr(destination);
-
-  if (!connect_ex(socket_, addr.addr(), static_cast<int>(addr.size), nullptr, 0,
-                  nullptr, reinterpret_cast<OVERLAPPED*>(&conn_overlapped_))) {
-    auto err_code = WSAGetLastError();
-    if (err_code != WSA_IO_PENDING) {
-      AE_TELED_ERROR("Socket connect error {}", err_code);
-      return ConnectionState::kConnectionFailed;
-    }
-  }
-  return ConnectionState::kConnecting;
-}
-
-WinTcpSocket::ConnectionState WinTcpSocket::GetConnectionState() {
-  int error = 0;
-  int opt_len = sizeof(error);
-  if (auto res = getsockopt(socket_, SOL_SOCKET, SO_ERROR,
-                            reinterpret_cast<char*>(&error), &opt_len);
-      res == SOCKET_ERROR) {
-    if (res == WSAEWOULDBLOCK) {
-      AE_TELED_DEBUG("Still connecting");
-      return ConnectionState::kConnecting;
-    }
-    AE_TELED_ERROR("Getsockopt returns error {}", res);
-    return ConnectionState::kConnectionFailed;
-  }
-
-  if (error != 0) {
-    if ((error == WSA_IO_PENDING) || (error == WSAEINPROGRESS)) {
-      AE_TELED_DEBUG("Still connecting");
-      return ConnectionState::kConnecting;
-    }
-    AE_TELED_ERROR("Socket connect error {}", error);
-    return ConnectionState::kConnectionFailed;
-  }
-
-  if (!connection_initiated_) {
-    if (connection_initiated_ = InitConnection(); !connection_initiated_) {
-      AE_TELED_ERROR("Failed to initiate connection");
-      return ConnectionState::kConnectionFailed;
-    }
   }
 
   return ConnectionState::kConnected;
 }
 
-void WinTcpSocket::Disconnect() {
+WinUdpSocket::ConnectionState WinUdpSocket::GetConnectionState() {
+  int error = 0;
+  int opt_len = sizeof(error);
+  if (auto res = getsockopt(socket_, SOL_SOCKET, SO_ERROR,
+                            reinterpret_cast<char*>(&error), &opt_len);
+      res == SOCKET_ERROR) {
+    AE_TELED_ERROR("Getsockopt returns error {}", res);
+    return ConnectionState::kConnectionFailed;
+  }
+
+  if (error != 0) {
+    AE_TELED_ERROR("Socket connect error {}", error);
+    return ConnectionState::kConnectionFailed;
+  }
+
+  return ConnectionState::kConnected;
+}
+
+void WinUdpSocket::Disconnect() {
   if (socket_ == INVALID_SOCKET) {
     return;
   }
   closesocket(socket_);
   socket_ = INVALID_SOCKET;
-}
-
-WinTcpSocket::ConnectExPtr WinTcpSocket::GetConnectEx() {
-  ConnectExPtr func_ptr = nullptr;
-  GUID guid = WSAID_CONNECTEX;
-  DWORD bytes;
-  int rc =
-      WSAIoctl(socket_, SIO_GET_EXTENSION_FUNCTION_POINTER, &guid, sizeof(guid),
-               &func_ptr, sizeof(func_ptr), &bytes, NULL, NULL);
-  if (rc == SOCKET_ERROR || func_ptr == nullptr) {
-    AE_TELED_ERROR("ConnectEx load failed: {}", WSAGetLastError());
-    return nullptr;
-  }
-  return func_ptr;
-}
-
-bool WinTcpSocket::InitConnection() {
-  // Update socket context for proper API usage
-  if (setsockopt(socket_, SOL_SOCKET, SO_UPDATE_CONNECT_CONTEXT, NULL, 0) ==
-      SOCKET_ERROR) {
-    AE_TELED_ERROR("SO_UPDATE_CONNECT_CONTEXT failed: {}", WSAGetLastError());
-    return false;
-  }
-
-  if (!RequestRecv()) {
-    AE_TELED_ERROR("Request receive failed!");
-    return false;
-  }
-  return true;
 }
 
 }  // namespace ae
